requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
pydantic
passlib[bcrypt]
python-jose
mysql-connector-python